feat(nav): unify mobile nav onto a collapsible sidebar drawer
Mobile (<lg) previously had two competing nav surfaces — a top section
tab strip plus a chat-only conversations drawer. Both are replaced by a
single collapsible drawer mirroring the desktop sidebar.

- Extract AppSidebar's inner content into a shared SidebarContent
  (brand + section nav + Conversations + footer); rendered by both the
  desktop <aside> and the mobile drawer.
- Add a section-aware MobileTopBar (mounted in layout.tsx): hamburger
  opens the sidebar mirror in a Sheet; center shows the conversation
  title + rename/delete menu on Chat, the section name elsewhere. Auth
  + About move to the drawer footer.
- Drop the visible TabsList from page.tsx (the Tabs container stays and
  still switches content panels). Remove the Chat mobile bar and the
  superseded ConversationsMobileSheet.
- Glossary: sharpen "tab" to canonical "Section" in CONTEXT.md.

Desktop is unchanged.
