+++ docs/progress.md
@@ -79,6 +79,8 @@ Multi-conversation, organised pantry, auth, and a leaner recipe surface. Highlig
 - **Shared dots→voice-lines loader** (Jun 2026): the chat (`ChatLoader`) and Tweak Bench (`TweakProgress`) share one loader vocabulary in `src/features/shared/components/loaders/` — `usePhaseAfter` (timed phase flip), `useCyclingIndex` + `CyclingVoiceLines`/`VoiceLine` (cross-fading Ah-Mah voice-lines), `shuffle`, and `LoadingDots`. Both show dots first, then — after **~2s** — an honest **indeterminate** indicator (chat's spinner ring / the Tweak Bench granny icon) beside shuffled, looping Ah-Mah voice-lines. **No progress bar**: there's no real signal to track — the tweak is a single buffered call (ADR-0010) and the chat loader only exists in the pre-render gap before the progressive recipe reveal (ADR-0009) takes over, so any bar would be a fake linear fill. Earlier iterations (segmented fill-and-hold, continuous easing bar) were dropped for this reason; the voice-lines shuffle per wait and cycle so a long wait stays lively instead of freezing.
 - **Retired** `SkeletonRecipeCard`, `WritingItOut`, and `ShimmerWord` (dead after skeleton removal). See ADR-0009.
 
+- **Mobile nav unified onto one collapsible drawer** (Jun 2026): mobile (`<lg`) previously had two competing nav surfaces — a top section tab strip (`page.tsx`'s `TabsList`) plus a chat-only conversations drawer. Both are gone. The desktop `AppSidebar`'s inner content is now a shared `SidebarContent` (brand + section nav + `Conversations` + footer) rendered by both the desktop `<aside>` and a mobile `Sheet`. A new section-aware `MobileTopBar` (mounted in `layout.tsx`) owns the drawer: ☰ on the left opens the full sidebar mirror; the center shows the active conversation title on Chat (with inline rename + the rename/delete `ConversationItemMenu`) and the section name ("Pantry"/"Cookbook") elsewhere. Auth + About now live only in the drawer footer. The `<Tabs>` container stays (it still switches the content panels); only the visible strip was dropped. `ConversationsMobileSheet` deleted. Glossary: "tab" sharpened to **Section** in `CONTEXT.md` (there is no visible tab strip anymore; the Radix Tabs are an implementation detail). Desktop unchanged.
+
 ### Vigilant pantry capture (gated pre-extraction) — Shipped Jun 2026
 - **Problem**: `gpt-4.1-mini` unreliably ignored the prompt rule to `addInventoryItem` before suggesting. "I have chicken broth, what can I cook?" often returned a recipe without ever adding the broth.
 - **Gated pre-extraction**: `captureMentionedInventory` (`src/lib/chat/captureInventory.ts`) runs server-side in `/api/chat` *before* `streamText`, so a subsequent `getInventory` already reflects mentioned items. A cheap regex gate (`mentionsPossession`) skips the extraction LLM call for pure questions; on a hit, a `temperature: 0` `generateObject` extracts only genuinely-possessed items (no negated/wished-for/dish-name items, no bare "with X" requests) and upserts them via `addInventoryItem` (idempotent on `userId_name_type`).
