+++ src/app/layout.tsx
@@ -1,11 +1,10 @@
-import { AuthButton } from "@/features/Auth";
 import { AppSidebar } from "@/features/shared/components/AppSidebar";
+import { MobileTopBar } from "@/features/shared/components/MobileTopBar";
 import { Toaster } from "@/components/ui/sonner";
 import { ConversationProvider } from "@/contexts/ConversationContext";
 import { SessionProvider } from "@/contexts/SessionContext";
 import type { Metadata } from "next";
 import { Fraunces, Inter, Nunito } from "next/font/google";
-import Image from "next/image";
 import { Suspense } from "react";
 import "./globals.css";
 
@@ -122,23 +121,9 @@ export default function RootLayout({
               <AppSidebar />
             </Suspense>
             <div className="flex-1 min-w-0 flex flex-col overflow-hidden">
-              {/* Mobile-only top bar */}
-              <div className="lg:hidden pb-2 pt-2 border-b px-4 flex justify-between items-center shrink-0">
-                <h1 className="text-xl font-display italic tracking-[-0.015em] leading-none flex items-center gap-2">
-                  <div className="relative w-8 h-8">
-                    <Image
-                      src="/granny-icon.png"
-                      alt="Ask Ah Mah"
-                      fill
-                      className="object-contain"
-                    />
-                  </div>
-                  <span className="font-normal text-ink-faint">Ask </span><span className="font-semibold text-primary">Ah Mah</span>
-                </h1>
-                <div className="flex items-center gap-2">
-                  <AuthButton />
-                </div>
-              </div>
+              <Suspense fallback={null}>
+                <MobileTopBar />
+              </Suspense>
               {children}
             </div>
           </ConversationProvider>
